:wrench: Renamed email to mail as the folder was confusing import
- Added http links to the footer and included the brand
- Added other locales for supported languages. These are currently in English while waiting for translation.
