+++ gabber/utils/email/locales/ar.json
