+++ gabber/utils/email/locales/es.json
@@ -1,6 +1,6 @@
 {
   "misc": {
-    "footer": "Download the {0} mobile application on Android ({1}) and iOS ({2}) to record and share your own conversations"
+    "footer": "Enregistrez et partagez votre propre conversation avec {0} sur Android et iOS."
   },
   "forgot": {
     "subject": "{0} restablecimiento de contraseña",
