+++ gabber/utils/email/locales/fr.json
